feat: :sparkles: storybook 7
