+++ packages/ripple-storybook/.storybook/utils/withTheme.js
@@ -1,23 +1,25 @@
-import themes from "../themes.js";
-import {provide} from 'vue'
+import themes from '../themes.js'
+import { provide } from 'vue'
 
 // Function to obtain the intended theme
 const getTheme = (themeName) => {
-  return Object.entries(themes[themeName]?.tokens || {}).reduce((result, [key, value]) => {
-    return {
-      ...result,
-      [`--${key}`]: value
-    }
-  }, {})
+  return Object.entries(themes[themeName]?.tokens || {}).reduce(
+    (result, [key, value]) => {
+      return {
+        ...result,
+        [`--${key}`]: value
+      }
+    },
+    {}
+  )
 }
 
-const withTheme=(story, context)=>{
-
+const withTheme = (story, context) => {
   return {
     components: { story },
     setup() {
-      const theme = getTheme(context.globals.theme);
-      const buttonTheme = context.globals.buttonTheme || 'default';
+      const theme = getTheme(context.globals.theme)
+      const buttonTheme = context.globals.buttonTheme || 'default'
 
       provide('featureFlags', {
         buttonTheme
@@ -25,8 +27,7 @@ const withTheme=(story, context)=>{
 
       return { theme }
     },
-    template:
-      `<div :style="theme">
+    template: `<div :style="theme">
         <story/>
       </div>`
   }
