+++ packages/ripple-storybook/package.json
@@ -6,38 +6,35 @@
   "license": "Apache-2.0",
   "private": true,
   "scripts": {
-    "start": "start-storybook -p 6006",
-    "build": "build-storybook",
-    "test": "test-storybook --stories-json",
-    "test:ci": "concurrently -k -s first -n \"SB,TEST\" -c \"magenta,blue\" \"npx http-server storybook-static --port 6006 --silent\" \"wait-on tcp:6006 && pnpm test\""
+    "start": "storybook dev -p 6006",
+    "static:build": "storybook build",
+    "static:start": "http-server storybook-static --port 6006 --silent",
+    "static:test": "wait-on tcp:6006 && test-storybook --stories-json --maxWorkers=2",
+    "test:ci": "concurrently -k -s first -n \"SB,TEST\" -c \"magenta,blue\" \"pnpm static:start\" \"pnpm static:test\""
   },
   "devDependencies": {
-    "@babel/core": "^7.21.3",
-    "@originjs/vite-plugin-content": "^1.0.3",
-    "@storybook/addon-a11y": "^6.5.16",
-    "@storybook/addon-actions": "^6.5.16",
-    "@storybook/addon-cssresources": "^6.2.9",
-    "@storybook/addon-essentials": "^6.5.16",
-    "@storybook/addon-interactions": "^6.5.16",
-    "@storybook/addon-jest": "^6.5.16",
-    "@storybook/builder-vite": "^0.2.7",
-    "@storybook/jest": "^0.0.10",
-    "@storybook/manager-webpack4": "^6.5.16",
-    "@storybook/test-runner": "^0.9.4",
-    "@storybook/testing-library": "^0.0.13",
-    "@storybook/vue3": "^6.5.16",
-    "@types/jest": "^29.5.0",
-    "@vitejs/plugin-vue": "^4.1.0",
-    "babel-loader": "^9.1.2",
-    "jest": "^29.5.0",
-    "jest-environment-jsdom": "^29.5.0",
-    "ts-jest": "^29.0.5",
-    "typescript": "^5.0.2",
-    "vite": "3.2.5",
-    "yaml": "^2.2.1"
+    "@storybook/addon-a11y": "^7.0.22",
+    "@storybook/addon-actions": "^7.0.22",
+    "@storybook/addon-essentials": "^7.0.22",
+    "@storybook/addon-interactions": "^7.0.22",
+    "@storybook/addon-jest": "^7.0.22",
+    "@storybook/addon-links": "^7.0.22",
+    "@storybook/blocks": "^7.0.22",
+    "@storybook/jest": "^0.1.1-future.2",
+    "@storybook/mdx1-csf": "^1.0.0",
+    "@storybook/test-runner": "^0.11.0-future.5",
+    "@storybook/testing-library": "^0.1.1-future.2",
+    "@storybook/vue3": "^7.0.22",
+    "@storybook/vue3-vite": "^7.0.22",
+    "concurrently": "^8.2.0",
+    "http-server": "^14.1.1",
+    "jest-axe": "^7.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "storybook": "^7.0.22",
+    "wait-on": "^7.0.1"
   },
   "dependencies": {
-    "concurrently": "^7.6.0",
     "vite-svg-loader": "^4.0.0"
   }
 }
